+++ packages/media/media-test-helpers/src/exampleMediaItems.ts
@@ -148,6 +148,18 @@ export const videoHorizontalFileId = {
   collectionName,
 };
 
+export const videoLargeFileId = {
+  id: '3291050e-6b66-4296-94c6-12088ef6fbad',
+  mediaItemType: fileType,
+  collectionName,
+};
+
+export const videoProcessingFailedId = {
+  id: 'e558199f-f982-4d23-93eb-313be5998d1b',
+  mediaItemType: fileType,
+  collectionName,
+};
+
 export const imageFileId = {
   id: '5556346b-b081-482b-bc4a-4faca8ecd2de', // jpg image
   mediaItemType: fileType,